// frontend/src/api/generated.d.ts
export interface components {
    schemas: {
        Health: {
            status: string;
        };
        Problem: {
            type: string;
            title: string;
            status: number;
            detail?: string;
            errors?: {
                [key: string]: string;
            };
        };
// ...
    };
// ...
}